refactor: make paragraph elements be reused
instead of clearing the canvas and regenerating new paragraph items every time a key is input or the buffer is invalidated. only regenerate the html when the window is resized, and iterate the html and insert the buffer row wise into the paragraphs. this should massively improve performance and will remove a huge headache in generalizing for more complex features in the future